improved benchmarking
