--- packages/job-components/bench/fixtures/proxy-entity.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const _metadata = new WeakMap();
-
-module.exports = function make(data, metadata) {
-    const proxy = new Proxy(data, {});
-    _metadata.set(proxy, Object.assign({ createdAt: Date.now() }, metadata));
-    return proxy;
-};
