+++ packages/utils/bench/data-encoding-suite.js
