+++ packages/utils/bench/fixtures/fake-data-entity.js
